useCollectionCreator for de-duplication.
