+++ client/src/components/Collections/PairCollectionCreator.vue
@@ -8,11 +8,10 @@ import type { HDASummary, HistoryItemSummary } from "@/api";
 import { useAnimationFrameResizeObserver } from "@/composables/sensors/animationFrameResizeObserver";
 import { useAnimationFrameScroll } from "@/composables/sensors/animationFrameScroll";
 import { Toast } from "@/composables/toast";
-import STATES from "@/mvc/dataset/states";
 import localize from "@/utils/localization";
 
 import type { DatasetPair } from "../History/adapters/buildCollectionModal";
-import { useExtensionFiltering } from "./common/useExtensionFilter";
+import { useCollectionCreator } from "./common/useCollectionCreator";
 import { guessNameForPair } from "./pairing";
 
 import GButton from "../BaseComponents/GButton.vue";
@@ -46,7 +45,6 @@ const emit = defineEmits<{
 }>();
 
 const state = ref("build");
-const removeExtensions = ref(true);
 const initialSuggestedName = ref(props.suggestedName);
 const invalidElements = ref<string[]>([]);
 const workingElements = ref<HDASummary[]>([]);
@@ -70,7 +68,6 @@ const noElementsSelected = computed(() => {
 const exactlyTwoValidElements = computed(() => {
     return pairElements.value.forward && pairElements.value.reverse;
 });
-const hideSourceItems = ref(props.defaultHideSourceItems || false);
 const pairElements = computed<SelectedDatasetPair>(() => {
     if (props.fromSelection) {
         return {
@@ -89,7 +86,7 @@ const pairHasMixedExtensions = computed(() => {
     );
 });
 
-const { hasInvalidExtension } = useExtensionFiltering(props);
+const { removeExtensions, hideSourceItems, onUpdateHideSourceItems, isElementInvalid } = useCollectionCreator(props);
 
 // check if we have scrolled to the top or bottom of the scrollable div
 const scrollableDiv = ref<HTMLDivElement | null>(null);
@@ -145,7 +142,7 @@ function _elementsSetUp() {
         const element = workingElements.value.find(
             (e) => e.id === inListElementsPrev[key as keyof SelectedDatasetPair]?.id
         );
-        const problem = _isElementInvalid(prevElem);
+        const problem = isElementInvalid(prevElem);
         if (element) {
             inListElements.value[key as keyof SelectedDatasetPair] = element;
         } else if (problem) {
@@ -180,7 +177,7 @@ function _ensureElementIds() {
 // /** separate working list into valid and invalid elements for this collection */
 function _validateElements() {
     workingElements.value = workingElements.value.filter((element) => {
-        var problem = _isElementInvalid(element);
+        var problem = isElementInvalid(element);
 
         if (problem) {
             invalidElements.value.push(element.name + "  " + problem);
@@ -190,29 +187,6 @@ function _validateElements() {
     });
 
     return workingElements.value;
-}
-
-/** describe what is wrong with a particular element if anything */
-function _isElementInvalid(element: HistoryItemSummary) {
-    if (element.history_content_type === "dataset_collection") {
-        return localize("is a collection, this is not allowed");
-    }
-
-    const validState = STATES.VALID_INPUT_STATES.includes(element.state as string);
-
-    if (!validState) {
-        return localize("has errored, is paused, or is not accessible");
-    }
-
-    if (element.deleted || element.purged) {
-        return localize("has been deleted or purged");
-    }
-
-    // is the element's extension not a subtype of any of the required extensions?
-    if (hasInvalidExtension(element)) {
-        return localize(`has an invalid extension: ${element.extension}`);
-    }
-    return null;
 }
 
 function getPairElement(key: string) {
@@ -259,7 +233,7 @@ function addUploadedFiles(files: HDASummary[]) {
     files.forEach((file) => {
         const element = workingElements.value.find((e) => e.id === file.id);
         if (element) {
-            const problem = _isElementInvalid(element);
+            const problem = isElementInvalid(file);
             if (problem) {
                 const invalidMsg = `${element.hid}: ${element.name} ${problem} and ${NOT_VALID_ELEMENT_MSG}`;
                 invalidElements.value.push(invalidMsg);
@@ -307,10 +281,6 @@ function removeExtensionsToggle() {
 function _guessNameForPair(fwd: HDASummary, rev: HDASummary, removeExtensions: boolean) {
     removeExtensions = removeExtensions ? removeExtensions : removeExtensions;
     return guessNameForPair(fwd, rev, "", "", removeExtensions);
-}
-
-function onUpdateHideSourceItems(newHideSourceItems: boolean) {
-    hideSourceItems.value = newHideSourceItems;
 }
 </script>
